Adds valid icon for HighResolutionIconUrl in mail
This commit adds an icon from the public domain (Tango UI library) and
adds it to the mail add-in generator so that the HighResolutionIconUrl
element in the manifest can refer to it. This will ensure that the
manifest file is valid and ready for installation at the very beginning,
reducing another pain point for developers.

Fix #48