Added option to pass command structure
Added option to pass predefined commands and functions for buttons

This is mainly to support customizing an existing sample, like the
'html' sample in the mail generator. The sample template can include an
XML stub with the Hosts and Resources nodes filled in for the manifest,
and a JS file with functions to add to the function file.