Fixing issues uncovered by tests
Testing revealed a couple of issues, this checkin fixes them.

- When running yo office and selecting Mail, HTML/CSS, if you select
CustomPane a CustomPane was not added to the manifest
- When running yo office and selecting Mail, HTML/CSS, if you unselect
any of the pre-selected extension points, it was added to the manifest
anyway.

Added a CustomPane template to the command data for the HTML template in
the mail generator.
Added a function to remove unused entries from the command data before
writing it to the manifest.