Fixing tests
- Moved call to `done` to the `on('end')` handler for a number of tests.
This resolves a timing issue, ensuring that all files are created before
testing if they exist.
- Added new `answer` value to all tests: `open: false`. This stops the
browser being opened multiple times to resource.html.
- Changed `folder` answer value for the `new project - answers & opts -
jquery & typescript` test group. The paths to the files were not
expecting them to be in a new subfolder.
- Removed the `options` variable in the `new project - answers & opts -
jquery & typescript` test group and instead declare them inline in each
sub-test. The values were perisisting, so the second test still had `js:
true` in the options passed, causing it to fail.
- Updated file list for the `'options: --js'` test. It was using the
list of files for a typescript project.