Bundle React by default.
Currently, the webpack configuration for the production build marks React and React-DOM as external. This means that OOB the add-in works in the development environment but when you go to production build it will result in error that React is undefined, because it is not included in the bundle, nor is it referenced explicitly with script links in the HTML page, This change will result in React being bundled by default. This will add about 3Mb to the bundle, but the client has to load this one way or another. You could explain in the documentation how to use a script link to CDN if that is the developer's preference.